Split oversized viewer modules
