+++ crates/pixel_world_bridge/src/lib.rs
@@ -17,6 +17,12 @@ use web_sys::HtmlCanvasElement;
 
 mod host_state;
 mod render;
+#[path = "lib_test_hit_targets.rs"]
+mod test_hit_targets;
+
+use test_hit_targets::{
+    hotspot_test_hit_targets, publish_hotspot_test_hit_targets, publish_location_test_hit_targets,
+};
 
 thread_local! {
     static BRIDGE_SHARED: RefCell<BridgeSharedState> = RefCell::new(BridgeSharedState::default());
@@ -892,37 +898,6 @@ fn hit_test(hit_regions: &[HitRegion], x: f64, y: f64) -> Option<(String, String
         }
     }
     None
-}
-
-fn hotspot_test_hit_targets(hit_regions: &[HitRegion]) -> Vec<HotspotTestHitTarget> {
-    hit_regions
-        .iter()
-        .filter(|region| region.kind == "hotspot")
-        .map(|region| HotspotTestHitTarget {
-            kind: region.kind,
-            id: region.id.clone(),
-            canvas_x: (region.left + region.right) / 2.0,
-            canvas_y: (region.top + region.bottom) / 2.0,
-        })
-        .collect()
-}
-
-fn publish_hotspot_test_hit_targets(hit_regions: &[HitRegion]) {
-    let targets = hotspot_test_hit_targets(hit_regions);
-    BRIDGE_SHARED.with(|shared| shared.borrow_mut().hotspot_test_targets = targets);
-}
-
-fn publish_location_test_hit_targets(hit_regions: &[HitRegion]) {
-    let targets = hit_regions
-        .iter()
-        .filter(|region| region.kind == "location")
-        .map(|region| LocationTestHitTarget {
-            id: region.id.clone(),
-            canvas_x: (region.left + region.right) / 2.0,
-            canvas_y: (region.top + region.bottom) / 2.0,
-        })
-        .collect();
-    BRIDGE_SHARED.with(|shared| shared.borrow_mut().location_test_targets = targets);
 }
 
 fn click_selection_from_hit(hit: Option<(String, String)>) -> Option<(String, String)> {
